rename Bio to StafCard
